Special update that adds a Pause Process button
Small stopwatch button will pause the RDR2 process for around 10 seconds to create a solo lobby. Requires the pause_server.py program to be running so it can send the message for windows to pause (browser security prevents it inline).
Will only appear locally, not on the github version for security reasons.

Also some updates to descriptions.